+++ Code/encoding/Structs/AnimationGraphDescriptorManager.h
@@ -9,21 +9,21 @@ struct AnimationGraphDescriptorManager
     TP_NOCOPYMOVE(AnimationGraphDescriptorManager);
 
     static AnimationGraphDescriptorManager& Get() noexcept;
-    const AnimationGraphDescriptor* GetDescriptor(size_t aKey) const noexcept;
+    const AnimationGraphDescriptor* GetDescriptor(uint64_t aKey) const noexcept;
 
     struct Builder
     {
-        Builder(AnimationGraphDescriptorManager& aManager, size_t aKey,
+        Builder(AnimationGraphDescriptorManager& aManager, uint64_t aKey,
                 AnimationGraphDescriptor aAnimationGraphDescriptor) noexcept;
     };
 
 protected:
 
-    void Register(size_t aKey, AnimationGraphDescriptor aAnimationGraphDescriptor) noexcept;
+    void Register(uint64_t aKey, AnimationGraphDescriptor aAnimationGraphDescriptor) noexcept;
 
 private:
 
     AnimationGraphDescriptorManager() noexcept;
 
-    Map<size_t, AnimationGraphDescriptor> m_descriptors;
+    Map<uint64_t, AnimationGraphDescriptor> m_descriptors;
 };
