+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_Bloodbug.cpp
@@ -92,7 +92,7 @@ enum Variables
 AnimationGraphDescriptor_Bloodbug::AnimationGraphDescriptor_Bloodbug(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 2494616162267432625;
+    uint64_t key = 2494616162267432625;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
