+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_CaveCricketRoot.cpp
@@ -50,7 +50,7 @@ enum Variables
 AnimationGraphDescriptor_CaveCricketRoot::AnimationGraphDescriptor_CaveCricketRoot(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 2852218639496266058;
+    uint64_t key = 2852218639496266058;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
