+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_Deer.cpp
@@ -122,7 +122,7 @@ enum Variables
 AnimationGraphDescriptor_Deer::AnimationGraphDescriptor_Deer(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 802028733968326397;
+    uint64_t key = 802028733968326397;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
