+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_Master_Behavior.cpp
@@ -306,7 +306,7 @@ enum Variables
 
 AnimationGraphDescriptor_Master_Behavior::AnimationGraphDescriptor_Master_Behavior(AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 6252079209869287675;
+    uint64_t key = 6252079209869287675;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
