+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_RabbitChicken.cpp
@@ -17,7 +17,7 @@ enum Variables
 AnimationGraphDescriptor_RabbitChicken::AnimationGraphDescriptor_RabbitChicken(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 2036755078842276651;
+    uint64_t key = 2036755078842276651;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
