+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_RootState.cpp
@@ -62,7 +62,7 @@ enum Variables
 AnimationGraphDescriptor_RootState::AnimationGraphDescriptor_RootState(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 5822185292726855210;
+    uint64_t key = 5822185292726855210;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
