+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_Stingwing.cpp
@@ -85,7 +85,7 @@ enum Variables
 AnimationGraphDescriptor_Stingwing::AnimationGraphDescriptor_Stingwing(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 7939498104135091671;
+    uint64_t key = 7939498104135091671;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
