+++ Code/encoding/Structs/Fallout4/AnimationGraphDescriptor_Wolflike.cpp
@@ -144,7 +144,7 @@ enum Variables
 AnimationGraphDescriptor_Wolflike::AnimationGraphDescriptor_Wolflike(
     AnimationGraphDescriptorManager& aManager)
 {
-    size_t key = 8042668663022388977;
+    uint64_t key = 8042668663022388977;
 
     AnimationGraphDescriptorManager::Builder s_builder(aManager, key,
         AnimationGraphDescriptor(
